feat(frontend): setup Apollo Client with proper typings
Closes #9
