+++ packages/frontend/modules/graphql/initialize-apollo.ts
@@ -1,9 +1,12 @@
 import { NormalizedCacheObject, ApolloClient } from 'apollo-boost'
-import createApolloClient from './createApolloClient'
 
-let apolloClient: ApolloClient<any> | null = null
+import createApolloClient from './create-apollo-client'
 
-export default function initializeApollo(initialState: NormalizedCacheObject): ApolloClient<any> {
+let apolloClient: ApolloClient<NormalizedCacheObject> | null = null
+
+export default function initializeApollo(
+  initialState: NormalizedCacheObject
+): ApolloClient<NormalizedCacheObject> {
   if (typeof window === 'undefined') {
     return createApolloClient(initialState)
   }
